Refactor: move action to separate module
